refactor: move reporter exports to `./reporters/*` and rename `reporter` option to `reporters`
Restructure subpath exports from `./node-reporter` and `./dev-reporter` to
`./reporters/node` and `./reporters/dev` for consistency with `./formatters/*`.
Rename `reporter` option in `CreateLoggerOptions` to `reporters` to match its
array-accepting semantics. Updated all source, tests, docs, and skills.
