+++ playground-lib/src/logger.ts
@@ -1,11 +1,11 @@
 import { consoleReporter, createLogger } from 'logs-sdk'
 // TODO: figure out a way to automatically add this reporter in the correct context
-import { devReporter } from 'logs-sdk/dev-reporter'
+import { devReporter } from 'logs-sdk/reporters/dev'
 import { diagnostics } from './diagnostics'
 
 export const log = createLogger({
   diagnostics: [diagnostics],
-  reporter: [consoleReporter, devReporter],
+  reporters: [consoleReporter, devReporter],
 })
 
 // TODO: explore this idea
